+++ packages/components-native/src/index.ts
@@ -17,6 +17,7 @@ export * from "./Divider";
 export * from "./EmptyState";
 export * from "./ErrorMessageWrapper";
 export * from "./Flex";
+export * from "./FormatFile";
 export * from "./Form";
 export * from "./FormField";
 export * from "./Heading";